requirements.txt
# LLMRank.io Backend Dependencies
# Core API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database & Storage
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI & LLM Integration
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0

# Web Scraping & Content Processing
trafilatura==1.6.4
requests==2.31.0
aiofiles==23.2.1

# Data Processing & Analysis
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# Monitoring & Scheduling
apscheduler==3.10.4
redis==5.0.1

# Utilities
tabulate==0.9.0
jsonschema==4.20.0
email-validator==2.1.0
python-multipart==0.0.6

# Security & Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Production Server
uvloop==0.19.0
gunicorn==21.2.0